requirements.txt
aioice==0.9.0
aiortc==1.9.0
altair==4.2.2
annotated-types==0.7.0
asttokens==2.4.1
attrs==24.2.0
audio-recorder-streamlit==0.0.10
av==12.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
breq==0.7
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-crc32c==1.5.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
gTTS==2.5.2
httplib2==0.22.0
idna==3.7
ifaddr==0.2.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
Js2Py==0.74
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pipwin==0.5.2
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.18.0
pyjsparser==2.7.1
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SpeechRecognition==3.10.4
stack-data==0.6.3
streamlit==1.37.1
streamlit-chat==0.1.1
streamlit-webrtc==0.47.7
tenacity==8.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
